+++ crates/holochain/tests/websocket.rs
@@ -1,17 +1,13 @@
 use ::fixt::prelude::*;
 use anyhow::Result;
-use assert_cmd::prelude::*;
 use futures::future;
-use futures::Future;
 use hdk::prelude::RemoteSignal;
 use holochain::sweettest::SweetAgents;
 use holochain::sweettest::SweetConductorBatch;
 use holochain::sweettest::SweetDnaFile;
 use holochain::{
-    conductor::api::ZomeCall,
     conductor::{
-        api::{AdminRequest, AdminResponse, AppRequest, AppResponse},
-        config::*,
+        api::{AdminRequest, AdminResponse},
         error::ConductorError,
         Conductor,
     },
@@ -25,88 +21,16 @@ use holochain_wasm_test_utils::TestWasm;
 use holochain_websocket::*;
 use matches::assert_matches;
 use observability;
-use std::path::PathBuf;
-use std::process::Stdio;
 use std::sync::Arc;
 use std::time::Duration;
 use tempdir::TempDir;
-use tokio::io::AsyncBufReadExt;
-use tokio::io::BufReader;
-use tokio::process::Child;
-use tokio::process::Command;
 use tokio_stream::StreamExt;
 use tracing::*;
 use url2::prelude::*;
 
 use test_utils::*;
 
-mod test_utils;
-
-pub fn spawn_output(holochain: &mut Child) {
-    let stdout = holochain.stdout.take();
-    let stderr = holochain.stderr.take();
-    tokio::task::spawn(async move {
-        if let Some(stdout) = stdout {
-            let mut reader = BufReader::new(stdout).lines();
-            while let Ok(Some(line)) = reader.next_line().await {
-                trace!("holochain bin stdout: {}", line);
-            }
-        }
-    });
-    tokio::task::spawn(async move {
-        if let Some(stderr) = stderr {
-            let mut reader = BufReader::new(stderr).lines();
-            while let Ok(Some(line)) = reader.next_line().await {
-                trace!("holochain bin stderr: {}", line);
-            }
-        }
-    });
-}
-
-pub async fn check_started(holochain: &mut Child) {
-    let started = tokio::time::timeout(std::time::Duration::from_secs(1), holochain.wait()).await;
-    if let Ok(status) = started {
-        panic!("Holochain failed to start. status: {:?}", status);
-    }
-}
-
-fn create_config(port: u16, environment_path: PathBuf) -> ConductorConfig {
-    ConductorConfig {
-        admin_interfaces: Some(vec![AdminInterfaceConfig {
-            driver: InterfaceDriver::Websocket { port },
-        }]),
-        environment_path: environment_path.into(),
-        network: None,
-        dpki: None,
-        passphrase_service: Some(PassphraseServiceConfig::FromConfig {
-            passphrase: "password".into(),
-        }),
-        keystore_path: None,
-        use_dangerous_test_keystore: true,
-    }
-}
-
-pub fn write_config(mut path: PathBuf, config: &ConductorConfig) -> PathBuf {
-    path.push("conductor_config.yml");
-    std::fs::write(path.clone(), serde_yaml::to_string(&config).unwrap()).unwrap();
-    path
-}
-
-#[instrument(skip(holochain, response))]
-async fn check_timeout<T>(
-    holochain: &mut Child,
-    response: impl Future<Output = Result<T, WebsocketError>>,
-    timeout_millis: u64,
-) -> T {
-    match tokio::time::timeout(std::time::Duration::from_millis(timeout_millis), response).await {
-        Ok(response) => response.unwrap(),
-        Err(_) => {
-            holochain.kill().await.unwrap();
-            error!("Timeout");
-            panic!("Timed out on request after {}", timeout_millis);
-        }
-    }
-}
+pub mod test_utils;
 
 #[tokio::test(flavor = "multi_thread")]
 #[cfg(feature = "slow_tests")]
@@ -130,18 +54,7 @@ async fn call_admin() {
         vec![(TestWasm::Foo.into(), TestWasm::Foo.into())],
     );
 
-    let cmd = std::process::Command::cargo_bin("holochain").unwrap();
-    let mut cmd = Command::from(cmd);
-    cmd.arg("--structured")
-        .arg("--config-path")
-        .arg(config_path)
-        .env("RUST_LOG", "debug")
-        .stdout(Stdio::piped())
-        .stderr(Stdio::piped())
-        .kill_on_drop(true);
-    let mut holochain = cmd.spawn().expect("Failed to spawn holochain");
-    spawn_output(&mut holochain);
-    check_started(&mut holochain).await;
+    let _holochain = start_holochain(config_path.clone()).await;
 
     let (mut client, _) = websocket_client_by_port(port).await.unwrap();
 
@@ -163,20 +76,20 @@ how_many: 42
 
     let orig_dna_hash = dna.dna_hash().clone();
     register_and_install_dna(
-        &mut holochain,
         &mut client,
         orig_dna_hash,
         fake_agent_pubkey_1(),
         fake_dna_path,
         Some(properties.clone()),
         "nick".into(),
+        6000,
     )
     .await;
 
     // List Dnas
     let request = AdminRequest::ListDnas;
     let response = client.request(request);
-    let response = check_timeout(&mut holochain, response, 6000).await;
+    let response = check_timeout(response, 6000).await;
 
     let tmp_wasm = dna.code().values().cloned().collect::<Vec<_>>();
     let mut tmp_dna = dna.dna_def().clone();
@@ -189,144 +102,6 @@ how_many: 42
 
     let expects = vec![dna.dna_hash().clone()];
     assert_matches!(response, AdminResponse::DnasListed(a) if a == expects);
-
-    holochain.kill().await.expect("Failed to kill holochain");
-}
-
-pub async fn start_holochain(config_path: PathBuf) -> Child {
-    tracing::info!("\n\n----\nstarting holochain\n----\n\n");
-    let cmd = std::process::Command::cargo_bin("holochain").unwrap();
-    let mut cmd = Command::from(cmd);
-    cmd.arg("--structured")
-        .arg("--config-path")
-        .arg(config_path)
-        .env("RUST_LOG", "trace")
-        .stdout(Stdio::piped())
-        .stderr(Stdio::piped())
-        .kill_on_drop(true);
-    let mut child = cmd.spawn().expect("Failed to spawn holochain");
-    spawn_output(&mut child);
-    check_started(&mut child).await;
-    child
-}
-
-pub async fn call_foo_fn(app_port: u16, original_dna_hash: DnaHash, holochain: &mut Child) {
-    // Connect to App Interface
-    let (mut app_tx, _) = websocket_client_by_port(app_port).await.unwrap();
-    let cell_id = CellId::from((original_dna_hash, fake_agent_pubkey_1()));
-    call_zome_fn(
-        holochain,
-        &mut app_tx,
-        cell_id,
-        TestWasm::Foo,
-        "foo".into(),
-        (),
-    )
-    .await;
-}
-
-pub async fn call_zome_fn<S>(
-    holochain: &mut Child,
-    app_tx: &mut WebsocketSender,
-    cell_id: CellId,
-    wasm: TestWasm,
-    fn_name: String,
-    input: S,
-) where
-    S: Serialize + std::fmt::Debug,
-{
-    let call: ZomeCall = ZomeCallInvocationFixturator::new(NamedInvocation(
-        cell_id,
-        wasm,
-        fn_name,
-        ExternIO::encode(input).unwrap(),
-    ))
-    .next()
-    .unwrap()
-    .into();
-    let request = AppRequest::ZomeCallInvocation(Box::new(call));
-    let response = app_tx.request(request);
-    let call_response = check_timeout(holochain, response, 6000).await;
-    trace!(?call_response);
-    assert_matches!(call_response, AppResponse::ZomeCallInvocation(_));
-}
-
-pub async fn attach_app_interface(
-    client: &mut WebsocketSender,
-    holochain: &mut Child,
-    port: Option<u16>,
-) -> u16 {
-    let request = AdminRequest::AttachAppInterface { port };
-    let response = client.request(request);
-    let response = check_timeout(holochain, response, 3000).await;
-    match response {
-        AdminResponse::AppInterfaceAttached { port } => port,
-        _ => panic!("Attach app interface failed: {:?}", response),
-    }
-}
-
-pub async fn retry_admin_interface(
-    port: u16,
-    mut attempts: usize,
-    delay: Duration,
-) -> WebsocketSender {
-    loop {
-        match websocket_client_by_port(port).await {
-            Ok(c) => return c.0,
-            Err(e) => {
-                attempts -= 1;
-                if attempts == 0 {
-                    panic!("Failed to join admin interface");
-                }
-                warn!(
-                    "Failed with {:?} to open admin interface, trying {} more times",
-                    e, attempts
-                );
-                tokio::time::sleep(delay).await;
-            }
-        }
-    }
-}
-
-async fn register_and_install_dna(
-    mut holochain: &mut Child,
-    client: &mut WebsocketSender,
-    orig_dna_hash: DnaHash,
-    agent_key: AgentPubKey,
-    dna_path: PathBuf,
-    properties: Option<YamlProperties>,
-    nick: String,
-) -> DnaHash {
-    let register_payload = RegisterDnaPayload {
-        uid: None,
-        properties,
-        source: DnaSource::Path(dna_path),
-    };
-    let request = AdminRequest::RegisterDna(Box::new(register_payload));
-    let response = client.request(request);
-    let response = check_timeout(&mut holochain, response, 6000).await;
-    assert_matches!(response, AdminResponse::DnaRegistered(_));
-    let dna_hash = if let AdminResponse::DnaRegistered(h) = response {
-        h.clone()
-    } else {
-        orig_dna_hash
-    };
-
-    let dna_payload = InstallAppDnaPayload {
-        hash: dna_hash.clone(),
-        nick,
-        membrane_proof: None,
-    };
-    let payload = InstallAppPayload {
-        dnas: vec![dna_payload],
-        installed_app_id: "test".to_string(),
-        agent_key,
-    };
-    let request = AdminRequest::InstallApp(Box::new(payload));
-    let response = client.request(request);
-    let response = check_timeout(&mut holochain, response, 6000).await;
-    assert_matches!(response, AdminResponse::AppInstalled(_));
-    dna_hash
 }
 
 #[tokio::test(flavor = "multi_thread")]
@@ -346,7 +121,7 @@ async fn call_zome() {
     let config = create_config(admin_port, environment_path);
     let config_path = write_config(path, &config);
 
-    let mut holochain = start_holochain(config_path.clone()).await;
+    let holochain = start_holochain(config_path.clone()).await;
 
     let (mut client, _) = websocket_client_by_port(admin_port).await.unwrap();
     let (_, receiver2) = websocket_client_by_port(admin_port).await.unwrap();
@@ -361,20 +136,20 @@ async fn call_zome() {
     // Install Dna
     let (fake_dna_path, _tmpdir) = write_fake_dna_file(dna.clone()).await.unwrap();
     let _dna_hash = register_and_install_dna(
-        &mut holochain,
         &mut client,
         original_dna_hash.clone(),
         fake_agent_pubkey_1(),
         fake_dna_path,
         None,
         "".into(),
+        6000,
     )
     .await;
 
     // List Dnas
     let request = AdminRequest::ListDnas;
     let response = client.request(request);
-    let response = check_timeout(&mut holochain, response, 3000).await;
+    let response = check_timeout(response, 3000).await;
 
     let expects = vec![original_dna_hash.clone()];
     assert_matches!(response, AdminResponse::DnasListed(a) if a == expects);
@@ -384,16 +159,16 @@ async fn call_zome() {
         installed_app_id: "test".to_string(),
     };
     let response = client.request(request);
-    let response = check_timeout(&mut holochain, response, 3000).await;
+    let response = check_timeout(response, 3000).await;
     assert_matches!(response, AdminResponse::AppEnabled { .. });
 
     // Attach App Interface
-    let app_port_rcvd = attach_app_interface(&mut client, &mut holochain, Some(app_port)).await;
+    let app_port_rcvd = attach_app_interface(&mut client, Some(app_port)).await;
     assert_eq!(app_port, app_port_rcvd);
 
     // Call Zome
     tracing::info!("Calling zome");
-    call_foo_fn(app_port, original_dna_hash.clone(), &mut holochain).await;
+    call_foo_fn(app_port, original_dna_hash.clone()).await;
 
     // Ensure that the other client does not receive any messages, i.e. that
     // responses are not broadcast to all connected clients, only the one
@@ -406,28 +181,25 @@ async fn call_zome() {
         .is_err());
 
     // Shutdown holochain
-    holochain.kill().await.expect("Failed to kill holochain");
+    std::mem::drop(holochain);
     std::mem::drop(client);
 
     // Call zome after restart
     tracing::info!("Restarting conductor");
-    let mut holochain = start_holochain(config_path).await;
+    let _holochain = start_holochain(config_path).await;
 
     tokio::time::sleep(std::time::Duration::from_millis(1000)).await;
 
     // Call Zome again on the existing app interface port
     tracing::info!("Calling zome again");
-    call_foo_fn(app_port, original_dna_hash, &mut holochain).await;
-
-    // Shutdown holochain
-    holochain.kill().await.expect("Failed to kill holochain");
+    call_foo_fn(app_port, original_dna_hash).await;
 }
 
 #[tokio::test(flavor = "multi_thread")]
 #[cfg(feature = "slow_tests")]
 async fn remote_signals() -> anyhow::Result<()> {
     observability::test_run().ok();
-    const NUM_CONDUCTORS: usize = 5;
+    const NUM_CONDUCTORS: usize = 2;
 
     let mut conductors = SweetConductorBatch::from_standard_config(NUM_CONDUCTORS).await;
 
@@ -488,17 +260,17 @@ async fn emit_signals() {
     // actually runs the holochain binary
 
     // TODO: B-01453: can we make this port 0 and find out the dynamic port later?
-    let port = 9911;
+    let admin_port = 9911;
 
-    let tmp_dir = TempDir::new("conductor_cfg_3").unwrap();
+    let tmp_dir = TempDir::new("conductor_cfg_emit_signals").unwrap();
     let path = tmp_dir.path().to_path_buf();
     let environment_path = path.clone();
-    let config = create_config(port, environment_path);
+    let config = create_config(admin_port, environment_path);
     let config_path = write_config(path, &config);
 
-    let mut holochain = start_holochain(config_path.clone()).await;
+    let _holochain = start_holochain(config_path.clone()).await;
 
-    let (mut admin_tx, _) = websocket_client_by_port(port).await.unwrap();
+    let (mut admin_tx, _) = websocket_client_by_port(admin_port).await.unwrap();
 
     let uuid = uuid::Uuid::new_v4();
     let dna = fake_dna_zomes(
@@ -511,13 +283,13 @@ async fn emit_signals() {
     let agent_key = fake_agent_pubkey_1();
 
     let dna_hash = register_and_install_dna(
-        &mut holochain,
         &mut admin_tx,
         orig_dna_hash,
         fake_agent_pubkey_1(),
         fake_dna_path,
         None,
         "".into(),
+        6000,
     )
     .await;
     let cell_id = CellId::new(dna_hash.clone(), agent_key.clone());
@@ -527,11 +299,11 @@ async fn emit_signals() {
         installed_app_id: "test".to_string(),
     };
     let response = admin_tx.request(request);
-    let response = check_timeout(&mut holochain, response, 3000).await;
+    let response = check_timeout(response, 3000).await;
     assert_matches!(response, AdminResponse::AppEnabled { .. });
 
     // Attach App Interface
-    let app_port = attach_app_interface(&mut admin_tx, &mut holochain, None).await;
+    let app_port = attach_app_interface(&mut admin_tx, None).await;
 
     ///////////////////////////////////////////////////////
     // Emit signals (the real test!)
@@ -540,7 +312,6 @@ async fn emit_signals() {
     let (_, app_rx_2) = websocket_client_by_port(app_port).await.unwrap();
 
     call_zome_fn(
-        &mut holochain,
         &mut app_tx_1,
         cell_id.clone(),
         TestWasm::EmitSignal,
@@ -570,9 +341,6 @@ async fn emit_signals() {
     assert_eq!(sig1, sig2);
 
     ///////////////////////////////////////////////////////
-
-    // Shutdown holochain
-    holochain.kill().await.expect("Failed to kill holochain");
 }
 
 #[tokio::test(flavor = "multi_thread")]
@@ -693,4 +461,80 @@ async fn too_many_open() {
         .unwrap();
     }
     conductor_handle.shutdown().await;
+}
+
+#[tokio::test(flavor = "multi_thread")]
+#[cfg(feature = "slow_tests")]
+// TODO: duplicate/rewrite this to also test happ bundles in addition to dna
+async fn concurrent_install_dna() {
+    use futures::StreamExt;
+
+    static NUM_DNA: u8 = 50;
+    static NUM_CONCURRENT_INSTALLS: u8 = 10;
+    static REQ_TIMEOUT_MS: u64 = 15000;
+
+    observability::test_run().ok();
+    // NOTE: This is a full integration test that
+    // actually runs the holochain binary
+
+    // TODO: B-01453: can we make this port 0 and find out the dynamic port later?
+    let admin_port = 9912;
+    // let app_port = 9914;
+
+    let tmp_dir = TempDir::new("conductor_cfg_concurrent_install_dna").unwrap();
+    let path = tmp_dir.path().to_path_buf();
+    let environment_path = path.clone();
+    let config = create_config(admin_port, environment_path);
+    let config_path = write_config(path, &config);
+
+    let _holochain = start_holochain(config_path.clone()).await;
+
+    let (client, _) = websocket_client_by_port(admin_port).await.unwrap();
+
+    let before = std::time::Instant::now();
+
+    let install_tasks_stream = futures::stream::iter((0..NUM_DNA).into_iter().map(|i| {
+        let zomes = vec![(TestWasm::Foo.into(), TestWasm::Foo.into())];
+        let mut client = client.clone();
+        tokio::spawn(async move {
+            let nick = format!("fake_dna_{}", i);
+
+            // Install Dna
+            let dna = fake_dna_zomes_named(&uuid::Uuid::new_v4().to_string(), &nick, zomes.clone());
+            let original_dna_hash = dna.dna_hash().clone();
+            let (fake_dna_path, _tmpdir) = write_fake_dna_file(dna.clone()).await.unwrap();
+            let agent_key = generate_agent_pubkey(&mut client, REQ_TIMEOUT_MS).await;
+            println!("[{}] Agent pub key generated", i);
+
+            let dna_hash = register_and_install_dna_named(
+                &mut client,
+                original_dna_hash.clone(),
+                agent_key,
+                fake_dna_path.clone(),
+                None,
+                nick.clone(),
+                nick.clone(),
+                REQ_TIMEOUT_MS,
+            )
+            .await;
+
+            println!(
+                "[{}] installed dna with hash {} and name {}",
+                i, dna_hash, nick
+            );
+        })
+    }))
+    .buffer_unordered(NUM_CONCURRENT_INSTALLS.into());
+
+    let install_tasks = futures::StreamExt::collect::<Vec<_>>(install_tasks_stream);
+
+    for r in install_tasks.await {
+        r.unwrap();
+    }
+
+    println!(
+        "installed {} dna in {:?}",
+        NUM_CONCURRENT_INSTALLS,
+        before.elapsed()
+    );
 }