+++ crates/holochain_p2p/src/types/wire.rs
@@ -29,6 +29,7 @@ pub(crate) enum WireMessage {
     },
     Publish {
         request_validation_receipt: bool,
+        countersigning_session: bool,
         dht_hash: holo_hash::AnyDhtHash,
         ops: Vec<(holo_hash::DhtOpHash, holochain_types::dht_op::DhtOp)>,
     },
@@ -55,6 +56,9 @@ pub(crate) enum WireMessage {
     },
     GetValidationPackage {
         header_hash: HeaderHash,
+    },
+    CountersigningAuthorityResponse {
+        signed_headers: Vec<SignedHeader>,
     },
 }
 
@@ -83,11 +87,13 @@ impl WireMessage {
 
     pub fn publish(
         request_validation_receipt: bool,
+        countersigning_session: bool,
         dht_hash: holo_hash::AnyDhtHash,
         ops: Vec<(holo_hash::DhtOpHash, holochain_types::dht_op::DhtOp)>,
     ) -> WireMessage {
         Self::Publish {
             request_validation_receipt,
+            countersigning_session,
             dht_hash,
             ops,
         }
@@ -127,5 +133,9 @@ impl WireMessage {
     }
     pub fn get_validation_package(header_hash: HeaderHash) -> WireMessage {
         Self::GetValidationPackage { header_hash }
+    }
+
+    pub fn countersigning_authority_response(signed_headers: Vec<SignedHeader>) -> WireMessage {
+        Self::CountersigningAuthorityResponse { signed_headers }
     }
 }