+++ crates/holochain_types/src/test_utils.rs
@@ -17,13 +17,23 @@ struct FakeProperties {
 
 /// A fixture example dna for unit testing.
 pub fn fake_dna_file(uid: &str) -> DnaFile {
-    fake_dna_zomes(uid, vec![("test".into(), vec![].into())])
+    fake_dna_file_named(uid, "test")
+}
+
+/// A named dna for unit testing.
+pub fn fake_dna_file_named(uid: &str, name: &str) -> DnaFile {
+    fake_dna_zomes_named(uid, name, vec![(name.into(), vec![].into())])
 }
 
 /// A fixture example dna for unit testing.
 pub fn fake_dna_zomes(uid: &str, zomes: Vec<(ZomeName, DnaWasm)>) -> DnaFile {
+    fake_dna_zomes_named(uid, "test", zomes)
+}
+
+/// A named dna for unit testing.
+pub fn fake_dna_zomes_named(uid: &str, name: &str, zomes: Vec<(ZomeName, DnaWasm)>) -> DnaFile {
     let mut dna = DnaDef {
-        name: "test".to_string(),
+        name: name.to_string(),
         properties: YamlProperties::new(serde_yaml::from_str("p: hi").unwrap())
             .try_into()
             .unwrap(),
@@ -83,7 +93,7 @@ pub async fn fake_unique_element(
 ) -> anyhow::Result<(SignedHeaderHashed, EntryHashed)> {
     let content: SerializedBytes =
         UnsafeBytes::from(nanoid::nanoid!().as_bytes().to_owned()).into();
-    let entry = EntryHashed::from_content_sync(Entry::App(content.try_into().unwrap()));
+    let entry = Entry::App(content.try_into().unwrap()).into_hashed();
     let app_entry_type = AppEntryTypeFixturator::new(visibility).next().unwrap();
     let header_1 = Header::Create(Create {
         author: agent_key,
@@ -96,7 +106,7 @@ pub async fn fake_unique_element(
     });
 
     Ok((
-        SignedHeaderHashed::new(&keystore, HeaderHashed::from_content_sync(header_1)).await?,
+        SignedHeaderHashed::new(&keystore, header_1.into_hashed()).await?,
         entry,
     ))
 }